requirements.txt
aiogram>=3.0
aiosqlite>=0.19
APScheduler>=3.10
python-dateutil>=2.8
python-dotenv>=1.0
